+++ service/controller/resource/prometheusremotewrite/resource.go
@@ -2,6 +2,7 @@ package prometheusremotewrite
 
 import (
 	"context"
+	"strings"
 
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
@@ -22,12 +23,20 @@ type Config struct {
 	K8sClient        k8sclient.Interface
 	Logger           micrologger.Logger
 	PrometheusClient promclient.Interface
+
+	HTTPProxy  string
+	HTTPSProxy string
+	NoProxy    string
 }
 
 type Resource struct {
 	k8sClient        k8sclient.Interface
 	logger           micrologger.Logger
 	prometheusClient promclient.Interface
+
+	HTTPProxy  string
+	HTTPSProxy string
+	NoProxy    string
 }
 
 func New(config Config) (*Resource, error) {
@@ -35,6 +44,10 @@ func New(config Config) (*Resource, error) {
 		k8sClient:        config.K8sClient,
 		logger:           config.Logger,
 		prometheusClient: config.PrometheusClient,
+
+		HTTPProxy:  config.HTTPProxy,
+		HTTPSProxy: config.HTTPSProxy,
+		NoProxy:    config.NoProxy,
 	}
 
 	return r, nil
@@ -69,21 +82,30 @@ func fetchPrometheusList(ctx context.Context, r *Resource, rw *pmov1alpha1.Remot
 	return prometheusList, nil
 }
 
-func ensurePrometheusRemoteWrite(r pmov1alpha1.RemoteWrite, p promv1.Prometheus) (*promv1.Prometheus, bool) {
-	r.Spec.RemoteWrite.Name = r.GetName()
+func (r *Resource) ensurePrometheusRemoteWrite(rw pmov1alpha1.RemoteWrite, p promv1.Prometheus) (*promv1.Prometheus, bool) {
+	rw.Spec.RemoteWrite.Name = rw.GetName()
+
+	if !strings.Contains(r.NoProxy, rw.Spec.RemoteWrite.URL) {
+		if len(r.HTTPSProxy) > 0 {
+			rw.Spec.RemoteWrite.ProxyURL = r.HTTPSProxy
+		} else if len(r.HTTPProxy) > 0 {
+			rw.Spec.RemoteWrite.ProxyURL = r.HTTPProxy
+		}
+	}
+
 	if p.Spec.RemoteWrite != nil {
-		if rwIndex, ok := remoteWriteExists(r.GetName(), p.Spec.RemoteWrite); !ok { // item not found
-			p.Spec.RemoteWrite = append(p.Spec.RemoteWrite, r.Spec.RemoteWrite)
+		if rwIndex, ok := remoteWriteExists(rw.GetName(), p.Spec.RemoteWrite); !ok { // item not found
+			p.Spec.RemoteWrite = append(p.Spec.RemoteWrite, rw.Spec.RemoteWrite)
 			return &p, true
-		} else if !cmp.Equal(r.Spec.RemoteWrite, p.Spec.RemoteWrite[rwIndex]) { //  item found
-			p.Spec.RemoteWrite[rwIndex] = r.Spec.RemoteWrite
+		} else if !cmp.Equal(rw.Spec.RemoteWrite, p.Spec.RemoteWrite[rwIndex]) { //  item found
+			p.Spec.RemoteWrite[rwIndex] = rw.Spec.RemoteWrite
 			return &p, true
 		} else {
 			// no update!!
 			return &p, false
 		}
 	} else {
-		p.Spec.RemoteWrite = []promv1.RemoteWriteSpec{r.Spec.RemoteWrite}
+		p.Spec.RemoteWrite = []promv1.RemoteWriteSpec{rw.Spec.RemoteWrite}
 		return &p, true
 	}
 }